+++ app/layout.tsx
@@ -22,6 +22,7 @@ import { getLocalePreferenceBootstrapScript } from "@/i18n/locale-preference";
 
 const GADS_CONVERSION_ID = "AW-18054900065";
 const REDDIT_PIXEL_ID = process.env.NEXT_PUBLIC_REDDIT_PIXEL_ID;
+const X_PIXEL_ID = process.env.NEXT_PUBLIC_X_PIXEL_ID ?? "q27qa";
 
 const dmSans = DM_Sans({
   variable: "--font-dm-sans",
@@ -162,7 +163,7 @@ if(!getCookie(cookie)){var a={version:${ATTRIBUTION_VERSION},first_landing_url:l
           {`!function(e,t,n,s,u,a){e.twq||(s=e.twq=function(){s.exe?s.exe.apply(s,arguments):s.queue.push(arguments);
           },s.version='1.1',s.queue=[],u=t.createElement(n),u.async=!0,u.src='https://static.ads-twitter.com/uwt.js',
           a=t.getElementsByTagName(n)[0],a.parentNode.insertBefore(u,a))}(window,document,'script');
-          twq('config','q27qa');`}
+          twq('config',${JSON.stringify(X_PIXEL_ID)});`}
         </Script>
         <NextIntlClientProvider messages={messages}>
           <PostHogProvider bootstrapUser={bootstrapUser}>